Adding frame buffer and creating orthographic camera controller

+++ Imp/src/Imp/Renderer/Renderer.cpp
@@ -47,7 +47,7 @@ namespace Imp
 
 	}
 
-	uint32_t Renderer::GetFrame()
+	uint64_t Renderer::GetFrame()
 	{
 		return RenderCommand::GetFrame();
 	}
